Make the following changes:
,

Find this element:

  <SidebarContent>

with the text content:

    The Whispering PeaksDashboardCampaignsCreator ToolsMap MakerJournalDice RollerCombat Tracker

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_6f1b4fdd._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Remove the divider at the top. In the SidebarMenu give the menu items padding on the left and right. Remove the margin that the slotclones have. I just want some space on the left and right of menu items so when you hover the background isn't touching the sides