Make the following changes:
,

Find this element:

  <Card className="flex flex-col shadow-lg hover:shadow-xl transition-shadow duration-300">

with the text content:

    The Whispering PeaksActiveAn adventure into the mysterious mountains where ancient secrets lie. Edit Delete

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_3099535c._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

On the campaign card put the campaign banner image at the top as a 16:9 image with no padding so it hits the edges. The under that put the campaign name, then the summary, then under that put a Set Active toggle. When this toggle is turned on all the other toggles on other campaigns should be turned off and this campaign should be set as active. Then make a footer for the card that has a 1px gray divider above it that you only see when you hover over the card and that is where the edit and delete buttons are.