Make the following changes:
,

Find this element:

  <Card className="group flex flex-col shadow-lg hover:shadow-xl transition-shadow duration-300 rounded-lg overflow-hidden">

with the text content:

    The Whispering PeaksAn adventure into the mysterious mountains where ancient secrets lie.Set as Active Campaign EditDelete

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_3099535c._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

On the campaign cards put the set as active in the footer and move the edit and delete buttons. I want the edit and delete buttons to only appear on hover of the card, they will appear in the top right of the card with a shadow behind them over the campaign image.