Make the following changes:
,

Find this element:

  <SidebarMenu className="px-2 pt-2">

with the text content:

    CampaignsSearch

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_38fd1019._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_57e79a6d._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Move the search menu item into the footer section of the sidebar above Dark Mode. Remove Campaigns menu item and get rid of the sidebar menu and the divider after it. This should move the campaign switcher up into the top under the sidebar header.