Make the following changes:
,

Find this element:

  <DiceRollerTool>

with the text content:

    Dice Roller Deluxe5Rolled d6 (5) = 5 +0 NoneAdvantageDisadvantageD4D6D8D10D12D20D100Roll HistoryYour last 20 rolls.[data-radix-scroll-area-viewport]{scrollbar-width:none;-ms-overflow-style:none;-webkit-overflow-scrolling:touch;}[data-radix-scroll-area-viewport]::-webkit-scrollbar{display:none}Rolled d6 (5) = 510:20:39 AM

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_b083e716._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_b083e716._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Move the roll description that comes up after a roll into the roll number result box at the bottom under the number, that way the whole card doesn't get expanded when the text becomes visible. Also change the format so when d20 with advantage or  disadvantage is rolled, don't write out the words for that. Instead show both rolls in parentheses and if rolling with advantage bold the higher roll and make it green and if rolling with disadvantage bold the lower roll and make it red.