Make the following changes:
,

Find this element:

  <Primitive.div dir="ltr" data-orientation="horizontal" className="w-full flex-1 flex flex-col min-h-0">

with the text content:

    PartyDiceCombatDice Roller Deluxe0+0NoneAdvantageDisadvantageD4D6D8D10D12D20D100Roll HistoryYour last 20 rolls.No rolls yet.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_b083e716._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_b083e716._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_b083e716._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

In the sidebar, make the tool switcher at the top have a black background and white text when active instead of the white background it currently has. Change the combat icon to swords. At the top of the Dice roller section remove the title text. Then reorder it so that the dice buttons are first, then the modifier, then the advantage options, and then the result box