Make the following changes:
,

Find this element:

  <Primitive.div dir="ltr" data-orientation="horizontal" className="w-full flex-1 flex flex-col min-h-0">

with the text content:

    PartyDiceCombat D4D6D8D10D12D20D100+0NoneAdvantageDisadvantage0Roll HistoryYour last 20 rolls.No rolls yet.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_b083e716._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_b083e716._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_b083e716._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

For the Advantage toggle section remove the None option. Make the advantage and disadvantage buttons be inactive by default. If the user clicks Advantage turn it green and if Disadvantage turn that one red. Only one can be on at a time. Clicking them also turns them back off.