Make the following changes:
,

Find this element:

  <Primitive.div dir="ltr" data-orientation="horizontal" className="w-full flex-1 flex flex-col min-h-0">

with the text content:

    PartyDiceCombat D4D6D8D10D12D20D100+0AdvantageDisadvantage0Roll HistoryYour last 20 rolls.No rolls yet.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Change the hover color of the advantage disadvantage to not change the background but to change the border to green and red respectively. Also change the text on them to ADV and DIS. Put these two buttons one above the other in a column in the same row as the modifier element to the right.