Make the following changes:
,

Find this element:

  <Primitive.div dir="ltr" data-orientation="horizontal" className="w-full flex-1 flex flex-col min-h-0">

with the text content:

    PartyDiceCombat D4D6D8D10D12D20D100+0ADVDIS0Roll HistoryYour last 20 rolls.No rolls yet.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

For the advantage buttons: when hovered over change the background to green and red respectively. Change the text back to Advantage and Disadvantage. Put a little label text above the modifier that says Modifiers. Get rid of the divider between the modifiers section and the roll result box. Reduce the gap at the top between the dice buttons and the top of the card.