Make the following changes:
,

Find this element:

  <DiceRollerTool>

with the text content:

    D4D6D8D10D12D20D100Modifiers+0AdvantageDisadvantage0Roll HistoryYour last 20 rolls.No rolls yet.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Add a little bit more space to the top between the card top and the dice buttons. For the column with the modifiers text and modifier element, make it so the text is at the top of the the column and the modifier plus minus thing is at the bottom so it is more in-line with the advantage buttons. For the advantage buttons, make the green and red color that is on hover the full red and green colors not faded ones. When they are active, keep the green and red borders as well.