Make the following changes:
,

Find this element:

  <CardContent className="space-y-4 px-4 pt-4 pb-4">

with the text content:

    D4D6D8D10D12D20D100Coin FlipModifiers+0AdvantageDisadvantage0

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_551be34b._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Change the text of the coin flip button to 50/50. Add a button full width above the results box that says Roll Dice. We are going to change how the dice buttons work. Change it so that when you click a dice button it is added to the whole roll. Example: click d4 three times and now in the results box it says 3d4. Then you click roll dice and it rolls all 3d4 and puts a result in the box. So you build the roll with the dice buttons then roll them all with the roll dice button. Do not include the 50/50 button in this logic. Give the 50/50 button a black border to differentiate it. In the box if it says 3d4 + 1d8 + 6 you should be able to click any of those to remove so clicking 3d4 would remove 1d4 from the roll. Clicking the +6 would remove the modifier from the roll.