Make the following changes:
,

Find this element:

  <span class="text-muted-foreground">

with the text content:

    Build your roll

Relevant files:

And change it as follows:

Make the build roll text of 3d4 + 1d6 +1 (example) bold and black. When they are hovered, make the background red and the text white. Make a tooltip that says "Remove Die" or "Remove Modifier" for the modifier hover.