Make the following changes:
,

Find this element:

  <CardContent className="space-y-4 px-4 pt-4 pb-4">

with the text content:

    D4D6D8D10D12D20D10050/50Modifiers+0AdvantageDisadvantageRoll Dice12[object Object] + [object Object] = 12

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_551be34b._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Change the text of the 50/50 button to Coin Flip, Coin above and Flip on the second line. In the results description I am seeing [object Object] + [object Object] + 2 + 2 = 7" instead of "d4(3) + d6(2) +2 = 7" which is how it should be written in the results box and also in the roll log.