Make the following changes:
,

Find this element:

  <DiceRollerTool>

with the text content:

    D4D6D8D10D12D20D100CoinFlipModifiers+0AdvantageDisadvantageRoll DiceBuild your rollRoll HistoryNo rolls yet.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Combine the roll history card with the dice roller card. remove the title text of roll history. change the layout so there is a 1px divider between the dice roll section and the roll history. The roll history should have a light gray background. It should just be a scrollable section of rows.