Make the following changes:
,

Find this element:

  <CardContent className="space-y-4 px-4 pt-4 pb-4">

with the text content:

    D4D6D8D10D12D20D100CoinFlipModifiers+0AdvantageDisadvantageRoll DiceBuild your rollNo rolls yet.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_551be34b._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Make the bottom of this card similar to the bottom of the campaign cards in the campaign manager. The bottom should have a light gray background and be full width with no padding. The divider between the top and bottom should be the 1px gray line. The bottom should be full width rows of past rolls and it should be scrollable but just the bottom section.