Make the following changes:
,

Find this element:

  <Primitive.div dir="ltr" data-orientation="horizontal" className="w-full flex-1 flex flex-col min-h-0">

with the text content:

    PartyDiceCombat  D4D6D8D10D12D20D100CoinFlipModifiers+0AdvantageDisadvantageRoll DiceBuild your rollNo rolls yet.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Make sure the div goes all the way to the bottom of the page so the card goes all the way to the bottom (with the little bit of padding similar to the right and left edges). Right now I see a medium gap between the bottom of the card and the bottom of the page.