Make the following changes:
,

Find this element:

  <DiceRollerTool>

with the text content:

    D4D6D8D10D12D20D100CoinFlipModifiers+0AdvantageDisadvantageRoll DiceBuild your rollNo rolls yet.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_19e04f56._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Add a small gap between the bottom of the dice builder/results box and the divider with the roll history. Remove the toast that pops up after a coin flip. Change the background and border of the coin flip button to yellow. Add a button next to the roll dice button that is called Quick d20 which just rolls 1d20 or 2 if the advantage/disadvantage is activated.