Make the following changes:
,

Find this element:

  <CardContent className="space-y-4 px-4 pt-4 pb-0">

with the text content:

    D4D6D8D10D12D20D100 FLIPModifiers+0AdvantageDisadvantageRoll DiceBuild your roll

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_551be34b._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_69288ac2._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_69288ac2._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Remove the icon from the Roll Dice button. Add a tooltip to the coin flip button that says "Heads or Tails". Add a tooltip to the Advantage button that says "Higher of 2d20" and "Lower of 2d20" for the disadvantage button. Remove the text Modifiers.