Make the following changes:
,

Find this element:

  <CardContent className="space-y-4 px-4 pt-4 pb-0">

with the text content:

    D4D6D8D10D12D20D100FLIP+0AdvantageDisadvantageRoll DiceBuild your roll

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_bcad1792._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_69288ac2._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_69288ac2._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Make the advantage and disadvantage buttons w-full and have them just fill the row with the modifier element all in one line. Change the text to ADV. and DIS. respectively. Change the background color of the coin flip button to be black and the icon and text to be white. Reduce the vertical space between the icon and text on the coin flip a little.