Make the following changes:
,

Find this element:

  <Button className="w-full" size="sm" variant="default">

with the text content:

    Roll Entire Party Initiative

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_65660148._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_e553ee34._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_e553ee34._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

When the roll party initiative is clicked, hide the button immediately and only have it come back when combat is ended. When it is clicked roll initiative for all party members with their initiative modifiers and then add them to the order with highest at the top. Assign a color to each player. Put the initiative number highlighted on the left of the tile. Add the player's AC and HP to the tile.