Make the following changes:
,

Find this element:

  <CombatTrackerTool>

with the text content:

    Add Player Add Enemy Add Party Initiative Order3ElaraHP:/ 28AC: 1216GromHP:/ 52AC: 15 Start Combat End Combat

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_6331acac._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_6331acac._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Add a tiny bit more of left and right side padding to the sides of the combatant list. Change the header Initiative Order to just Initiative. On the combatant tiles move the AC and HP to the bottom right and the trash can to the top right. For HP just use the heart icon and the #/# and then for DC just use the shield and the #.