Make the following changes:
,

Find this element:

  <CombatTrackerTool>

with the text content:

    Add Player Add Enemy Initiative(Round 1 - Elara's turn)22Elara28 / 281216Grom52 / 5215Next Turn  End Combat

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_6331acac._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_6331acac._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

In the combatant list change the round 1 text to just show the current round and not the current combatants turn. On the combatant tile, add Player under the player character name. Move the shield AC up to the left of the trash can symbol. For the trash can make the hover background red and the icon white.