Make the following changes:
,

Find this element:

  <CombatTrackerTool>

with the text content:

    Add Combatant  Add Party Initiative Order14ElaraPlayer28 / 28127GromPlayer52 / 5215 Start Combat End Combat

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

On the combatant tile replace the current HP with a health meter that has the hp #/# and a percentage. The bar should be black for the full part and gray for the empty part. The black turns red when it gets below 20%.