Make the following changes:
,

Find this element:

  <CombatTrackerTool>

with the text content:

    Add Combatant  Add Party Initiative Order5ElaraPlayer28 / 28100%1215GromPlayer52 / 52100%15 Start Combat End Combat

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

On the combatant tile remove the Player text and remove the trash can icon. Move the AC into the top right. When the combatant is clicked open a little popup with options to delete the combatant or add a hit/heal. This will be Delete option in one row and in the second row there will be a number entry field with two buttons next to it, one that is red that says Hit one that is green that says Heal.