Make the following changes:
,

Find this element:

  <CombatTrackerTool>

with the text content:

    Add Combatant  Add Party Initiative OrderAdd combatants to begin.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

in the initiative order header, add a history icon to the far right inline with the initiative order text. When this is clicked open a popup to the left of the sidebar that has an encounter history. This log is added to every time that combat is ended. So the end combat button adds the end result of player health and list of dead enemies/allies and puts it in the log and then clears the initiative order.