Make the following changes:
,

Find this element:

  <SidebarMenu className="px-2">

with the text content:

    Party ManagerEncountersFactionsMaps & BoardsRandom TablesQuestingSession JournalLore & World

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_36ab3727._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Combine questing and Factions menu items into one item called Factions & Quests. Make the page /factions-quests and create the page now. This page should be dedicated to tracking the different factions in the world that you can have reputation with. NPCs will belong to a faction or be Unaffiliated. Any NPC information sheet will display the party's reputation with that faction. The reputation is a scale of -20 to +20. There are 7 milestones on this scale: Hated,  Hostile,  Unfriendly,  Neutral,  Friendly,  Allied,  Exalted. These are represented at the point intervals -20, -10, -5, 0, +5, +10, and +20. Create some dummy factions and repuations for the campaigns based on their world.