Make the following changes:
,

Find this element:

  <DiceRollerTool>

with the text content:

    Build your rollD4D6D8D10D12D20D100FLIP+0ADV.DIS.Roll DiceNo rolls yet.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

For the Dice Roller, give it a header with dice icon called Dice Roller. On the right side in line with the header add a history icon. Move the roll history into a drawer that opens from the right side with a title of Roll History and segment the roll history list items by day.