Make the following changes:
,

Find this element:

  <DiceRollerTool>

with the text content:

    Dice RollerRoll HistoryBuild your rollD4D6D8D10D12D20D100FLIP+0ADV.DIS.Roll Dice Reset Dice

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Change the spinning icon so that when dice are being rolled it spins the dices icon and when coin flip is happening it spins the disc-3 icon like it does now. Make no other changes to any other element in the diceroller.