Make the following changes:
,

Find this element:

  <DiceRollerTool>

with the text content:

    Dice RollerRoll HistoryBuild your rollD4D6D8D10D12D20D100FLIP+0ADV.DIS.Roll Dice Reset DiceNo rolls yet.

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Remove the roll history from the bottom of the dice roller. It should only be in the Roll History sheet that comes in from the right. Make no other changes besides removing that from the bottom. Leave the Last Roll element.