Make the following changes:
,

Find this element:

  <DialogHeader className="p-4 sm:p-6 border-b">

with the text content:

    Elara

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_8d3cffdb._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Move the profile picture, the race/class/background etc and the experience bar all into the header. On the left should be the profile picture and next to it on the right is the three rows of that information. Make the profile picture only as tall as those three rows of text/experience bar. Then on the far right of the header put the New Portrait button oriented/aligned to the top.