Make the following changes:
,

Find this element:

  <Button variant="outline" size="sm">

with the text content:

    Edit

Relevant files:

    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/src_2a79205b._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_c6392419._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_c6392419._.js
    - https://6000-firebase-studio-1747165847076.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Change the edit button trigger. When this button is clicked I want this character dialog to become editable. All the items like Level, race, class, initiative, the ability scores, the feats, the backstory etc. Do not have this button trigger the Character Creator dialog like it used to. Do nothing else, do not make any other changes to the layout of the rest of the app.